Use @spyke_attributes instead of @attributes instance var
The natural choice of instance variable naming here would be
`@attributes`. Unfortunately that causes a naming clash when used with
`ActiveModel::Dirty` version >= 5.2.0.

As of v5.2.0 `ActiveModel::Dirty` checks to see if `ActiveRecord`
attributes exist, assuming that if the instance variable `@attributes`
exists on the instance, it is because they are `ActiveRecord`
attributes.